+++ packages/privacy/test/index.spec.ts
@@ -1,5 +1,0 @@
-import { describe, it } from "vitest";
-
-describe("dummy", () => {
-    it.skip("dummy", () => {});
-});