+++ examples/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @examples/ui
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [52861e5]
+  - @wonderland/interop-addresses@0.1.1
+
 ## 0.1.1
 
 ### Patch Changes